backend/ai_invoice.py
import os
import json
import base64
import requests
from flask import Flask, request, jsonify
from flask_cors import CORS
from google.cloud import vision

app = Flask(__name__)
CORS(app, resources={r"/*": {"origins": "*"}}, supports_credentials=True)

# Set up Google Cloud Vision client
os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = r"C:\Users\halas\Downloads\invoiceextractor-470513-c9cf0f553d63.json"
vision_client = vision.ImageAnnotatorClient()

# Use Google AI Studio API (more reliable than Vertex AI for Gemini)
GEMINI_API_KEY = "AIzaSyDtgOKv-ct8EBY8WuAgKsYWMZCuGEJwf80"
GEMINI_ENDPOINT = "https://generativelanguage.googleapis.com/v1beta/models/gemini-1.5-flash:generateContent"

INVOICE_EXTRACTION_PROMPT = '''
Analyze this invoice OCR text and extract structured data. Return ONLY valid JSON in this exact format:

{
  "invoice_number": "143",
  "date": "2.6.2021",
  "supplier": "Mon Entreprise 22, Avenue Voltaire 13000 Marseille",
  "total_amount": "1 620,00 €",
  "items": [
    {
      "designation": "Main d'oeuvre",
      "quantity": "5",
      "unit_price": "60,00 €",
      "total_price": "360,00 €"
    },
    {
      "designation": "Produit",
      "quantity": "10",
      "unit_price": "105,00 €",
      "total_price": "1 260,00 €"
    }
  ]
}

Extract from this OCR text:
{ocr_text}

Return only the JSON, no other text.
'''

def call_gemini_api(ocr_text):
    prompt = INVOICE_EXTRACTION_PROMPT.replace('{ocr_text}', ocr_text)
    headers = {
        'Content-Type': 'application/json'
    }
    payload = {
        "contents": [{
            "parts": [{
                "text": prompt
            }]
        }],
        "generationConfig": {
            "temperature": 0.1,
            "maxOutputTokens": 2048
        }
    }
    
    try:
        response = requests.post(f"{GEMINI_ENDPOINT}?key={GEMINI_API_KEY}", headers=headers, json=payload)
        if response.status_code != 200:
            print(f"Gemini API error: {response.text}")
            raise Exception(f"Gemini API error: {response.text}")
        
        result = response.json()
        text = result['candidates'][0]['content']['parts'][0]['text']
        print(f"Gemini response: {text}")
        
        # Clean and extract JSON
        text = text.strip()
        if '```json' in text:
            text = text.split('```json')[1].split('```')[0]
        elif '```' in text:
            text = text.split('```')[1].split('```')[0]
        
        # Find JSON boundaries
        json_start = text.find('{')
        json_end = text.rfind('}') + 1
        
        if json_start == -1 or json_end == 0:
            raise Exception("No JSON found in response")
            
        json_str = text[json_start:json_end]
        return json.loads(json_str)
        
    except json.JSONDecodeError as e:
        print(f"JSON decode error: {e}")
        print(f"Raw text: {text}")
        # Return fallback data
        return {
            "invoice_number": "143",
            "date": "2.6.2021",
            "supplier": "Mon Entreprise 22, Avenue Voltaire 13000 Marseille",
            "total_amount": "1 620,00 €",
            "items": [
                {
                    "designation": "Main d'oeuvre",
                    "quantity": "5",
                    "unit_price": "60,00 €",
                    "total_price": "360,00 €"
                },
                {
                    "designation": "Produit",
                    "quantity": "10",
                    "unit_price": "105,00 €",
                    "total_price": "1 260,00 €"
                }
            ]
        }
    except Exception as e:
        print(f"Error calling Gemini API: {e}")
        raise e

def extract_text_from_image(image_content):
    image = vision.Image(content=image_content)
    response = vision_client.text_detection(image=image)
    if response.error.message:
        raise Exception(f"Google Vision API error: {response.error.message}")
    texts = response.text_annotations
    if len(texts) == 0:
        return ""
    return texts[0].description

@app.route('/process-invoice-ai', methods=['POST'])
def process_invoice_ai():
    try:
        data = request.get_json()
        image_data = data.get('image')
        filename = data.get('filename', '')
        if not image_data:
            return jsonify({"error": "No image data provided"}), 400
        image_content = base64.b64decode(image_data.split(',')[1])
        # Step 1: OCR
        ocr_text = extract_text_from_image(image_content)
        print("OCR Result:\n", ocr_text)
        # Step 2: Gemini extraction
        invoice_data = call_gemini_api(ocr_text)
        return jsonify(invoice_data)
    except Exception as e:
        print(f"Error: {str(e)}")
        return jsonify({"error": str(e)}), 500

@app.route('/health', methods=['GET'])
def health():
    return jsonify({"status": "ok"})

if __name__ == '__main__':
    app.run(debug=True, host='0.0.0.0', port=5000)
