backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-cloud-vision==3.4.4
requests==2.31.0
Pillow==10.0.1
